requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.0.0
dataclasses==0.6
Django==3.2.15
django-crispy-forms==1.13.0
html5lib==1.1
idna==3.3
lxml==4.9.1
numpy==1.19.5
Pillow==8.4.0
pytz==2022.2.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.2
urllib3==1.26.11
webencodings==0.5.1
